backend/requirements.txt
# FastAPI 核心
fastapi==0.104.1
uvicorn[standard]==0.24.0

# 数据库相关
sqlalchemy==2.0.23
pymysql==1.1.0
cryptography==41.0.7

# 认证和安全
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
passlib[bcrypt]==1.7.4

# 配置管理
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0

# 邮箱验证
email-validator==2.1.0

# 开发工具
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2